refactor: create redis abstraction
